Check if port is already used for run Grasscutter

+++ gcandroid/runGrasscutter.sh
@@ -19,6 +19,16 @@ GoTouchGrass() {
         main_menu
         return
     else
+        echo "${GC}Checking port used${WC}"
+        checkPortUsed=$(timeout --foreground 1s php -S $(jq -r .server.http.accessAddress $HOME/Grasscutter/config.json):$(jq .server.http.bindPort $HOME/Grasscutter/config.json) &>/dev/null)
+        if [[ $? -eq 1 ]]; then
+            echo "${RC}Error${WC} : Port $(jq .server.http.bindPort $HOME/Grasscutter/config.json) is used/permission denied!"
+            echo -n "Press enter for back to Menu!"
+            read -r
+            main_menu
+            return
+        fi
+        echo "${GC}Port $(jq .server.http.bindPort $HOME/Grasscutter/config.json) is not used!${WC}"
         pkill mongod
         mongod &>/dev/null &
         cd $HOME/Grasscutter
